+++ src/types/request.rs
@@ -612,16 +612,29 @@ impl Default for Request {
 impl From<hyper::Request<Body>> for Request {
     #[inline]
     fn from(from: hyper::Request<Body>) -> Self {
-        let time = SystemTime::now();
         let (head, body) = from.into_parts();
-
         Self {
-            created: time,
+            created: SystemTime::now(),
             method: head.method,
             uri: head.uri,
             version: head.version,
             headers: head.headers,
             body,
+        }
+    }
+}
+
+/// Create a new Request based on hyper's Request
+impl From<hyper::http::request::Parts> for Request {
+    #[inline]
+    fn from(head: hyper::http::request::Parts) -> Self {
+        Self {
+            created: SystemTime::now(),
+            method: head.method,
+            uri: head.uri,
+            version: head.version,
+            headers: head.headers,
+            body: Body::empty(),
         }
     }
 }